[REF] make id of user integer instead of long

+++ users/src/main/java/be/vinci/chattycar/users/UsersRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsersRepository extends CrudRepository<User, Long> {
+public interface UsersRepository extends CrudRepository<User, Integer> {
 
   boolean existsByEmail(String email);
   User findByEmail(String email);
